use common method for announcing deploy start
